@@ -11,27 +11,26 @@
   },
   "author": "Hariharan",
   "dependencies": {
-    "@emailjs/browser": "^4.3.3",
-    "framer-motion": "^11.2.4",
-    "react": "^18.2.0",
+    "@emailjs/browser": "^4.4.1",
+    "framer-motion": "^12.27.1",
+    "react": "^19.2.3",
     "react-countup": "^6.5.3",
-    "react-dom": "^18.2.0",
-    "react-icons": "^5.2.1",
-    "react-scroll": "^1.9.0",
-    "react-scroll-trigger": "^0.6.14",
-    "react-vertical-timeline-component": "^3.6.0"
+    "react-dom": "^19.2.3",
+    "react-icons": "^5.5.0",
+    "react-vertical-timeline-component": "^3.5.3"
   },
   "devDependencies": {
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
-    "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.3",
-    "vite": "^5.2.0"
+    "@tailwindcss/postcss": "^4.1.18",
+    "@types/react": "^19.2.8",
+    "@types/react-dom": "^19.2.3",
+    "@vitejs/plugin-react": "^5.1.2",
+    "autoprefixer": "^10.4.23",
+    "eslint": "^9.39.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "eslint-plugin-react-refresh": "^0.4.26",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.18",
+    "vite": "^7.3.1"
   }
 }
