@@ -1,4 +1,3 @@
-import React, { Component } from "react";
 import { FaLinkedin, FaGithub, FaInstagram } from "react-icons/fa";
 import { FaCode } from "react-icons/fa6";
 import logo from "../assets/logo.png";
@@ -34,19 +33,19 @@ const SocialMediaLink = ({ Icon, url, label }) => (
   </a>
 );
 
-export default class Navbar extends Component {
-  render() {
-    return (
-      <nav className="mb-5 flex items-center justify-between py-6">
-        <div className="flex flex-shrink-0 items-center">
-          <img width={120} height={120} src={logo} alt="logo" />
-        </div>
-        <div className="m-8 flex items-center justify-center gap-4 text-2xl">
-          {socialLinks.map(({ Icon, url, label }) => (
-            <SocialMediaLink key={label} Icon={Icon} url={url} label={label} />
-          ))}
-        </div>
-      </nav>
-    );
-  }
-}
+const Navbar = () => {
+  return (
+    <nav className="mb-5 flex items-center justify-between py-6">
+      <div className="flex flex-shrink-0 items-center">
+        <img width={120} height={120} src={logo} alt="logo" />
+      </div>
+      <div className="m-8 flex items-center justify-center gap-4 text-2xl">
+        {socialLinks.map(({ Icon, url, label }) => (
+          <SocialMediaLink key={label} Icon={Icon} url={url} label={label} />
+        ))}
+      </div>
+    </nav>
+  );
+};
+
+export default Navbar;
