@@ -1,59 +1,58 @@
-import React, { Component } from "react";
 import { PROJECTS } from "../constants";
 import { motion } from "framer-motion";
 
-export default class Projects extends Component {
-  render() {
-    return (
-      <div className="border-b border-neutral-900 pb-8">
-        <motion.h1
-          whileInView={{ opacity: 1, y: 0 }}
-          initial={{ opacity: 0, y: -100 }}
-          transition={{ duration: 0.5 }}
-          className="my-20 text-center text-4xl"
-        >
-          Projects
-        </motion.h1>
-        <div>
-          {PROJECTS.map((project, index) => (
-            <div key={index} className="mb-8 flex flex-wrap lg:justify-center">
-              <motion.div
-                whileInView={{ opacity: 1, x: 0 }}
-                initial={{ opacity: 0, x: -100 }}
-                transition={{ duration: 1 }}
-                className="w-full lg:w-1/4"
-              >
-                <img
-                  width={150}
-                  height={150}
-                  src={project.image}
-                  alt={project.title}
-                  className="mb-6 rounded "
-                />
-              </motion.div>
-              <motion.div
-                whileInView={{ opacity: 1, x: 0 }}
-                initial={{ opacity: 0, x: -100 }}
-                transition={{ duration: 1 }}
-                className="lg:ml-6 w-full max-w-xl lg:w-3/4"
-              >
-                <h6 className="mb-2 font-semibold">{project.title}</h6>
-                <p className="mb-4 text-justify text-neutral-400">
-                  {project.description}
-                </p>
-                {project.technologies.map((tech, index) => (
-                  <span
-                    key={index}
-                    className="mr-2 rounded bg-neutral-900 px-2 py-1 text-sm font-medium text-green-700"
-                  >
-                    {tech}
-                  </span>
-                ))}
-              </motion.div>
-            </div>
-          ))}
-        </div>
+const Projects = () => {
+  return (
+    <div className="border-b border-neutral-900 pb-8">
+      <motion.h1
+        whileInView={{ opacity: 1, y: 0 }}
+        initial={{ opacity: 0, y: -100 }}
+        transition={{ duration: 0.5 }}
+        className="my-20 text-center text-4xl"
+      >
+        Projects
+      </motion.h1>
+      <div>
+        {PROJECTS.map((project, index) => (
+          <div key={index} className="mb-8 flex flex-wrap lg:justify-center">
+            <motion.div
+              whileInView={{ opacity: 1, x: 0 }}
+              initial={{ opacity: 0, x: -100 }}
+              transition={{ duration: 1 }}
+              className="w-full lg:w-1/4"
+            >
+              <img
+                width={150}
+                height={150}
+                src={project.image}
+                alt={project.title}
+                className="mb-6 rounded "
+              />
+            </motion.div>
+            <motion.div
+              whileInView={{ opacity: 1, x: 0 }}
+              initial={{ opacity: 0, x: -100 }}
+              transition={{ duration: 1 }}
+              className="lg:ml-6 w-full max-w-xl lg:w-3/4"
+            >
+              <h6 className="mb-2 font-semibold">{project.title}</h6>
+              <p className="mb-4 text-justify text-neutral-400">
+                {project.description}
+              </p>
+              {project.technologies.map((tech, index) => (
+                <span
+                  key={index}
+                  className="mr-2 rounded bg-neutral-900 px-2 py-1 text-sm font-medium text-green-700"
+                >
+                  {tech}
+                </span>
+              ))}
+            </motion.div>
+          </div>
+        ))}
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
+
+export default Projects;
